requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
Pillow>=10.0.0
python-dotenv>=1.0.0
